+++ bzt/modules/grinder.py
@@ -24,9 +24,10 @@ from bzt import TaurusConfigError, ToolError
 from bzt.engine import ScenarioExecutor, FileLister, HavingInstallableTools, SelfDiagnosable
 from bzt.modules.aggregator import ConsolidatingAggregator, ResultsReader
 from bzt.modules.console import WidgetProvider, ExecutorWidget
+from bzt.modules.java import TaurusJavaHelper
 from bzt.requests_model import HTTPRequest
 from bzt.six import iteritems
-from bzt.utils import shell_exec, MirrorsManager, dehumanize_time, get_full_path, PythonGenerator, TaurusJavaHelperJar
+from bzt.utils import shell_exec, MirrorsManager, dehumanize_time, get_full_path, PythonGenerator
 from bzt.utils import unzip, RequiredTool, JavaVM, shutdown_process, TclLibrary, FileReader
 
 
@@ -158,7 +159,7 @@ class GrinderExecutor(ScenarioExecutor, WidgetProvider, FileLister, HavingInstal
         # add logback configurations used by worker processes (logback-worker.xml)
         res_dir = os.path.join(get_full_path(__file__, step_up=2), 'resources')
         self.env.add_path({"CLASSPATH": res_dir}, finish=True)
-        self.env.add_path({"CLASSPATH": TaurusJavaHelperJar(self.log).tool_path}, finish=True)
+        self.env.add_path({"CLASSPATH": TaurusJavaHelper().tool_path}, finish=True)
         self.env.add_path({"CLASSPATH": self.settings.get("path", None)}, finish=True)
 
         self.cmd_line = ["java", "net.grinder.Grinder", self.properties_file]
@@ -234,7 +235,7 @@ class GrinderExecutor(ScenarioExecutor, WidgetProvider, FileLister, HavingInstal
         download_link = self.settings.get("download-link", "")
         required_tools = [TclLibrary(self.log),
                           JavaVM(self.log),
-                          TaurusJavaHelperJar(self.log),
+                          TaurusJavaHelper(),
                           Grinder(grinder_path, self.log, GrinderExecutor.VERSION, download_link=download_link)]
 
         for tool in required_tools:
@@ -422,7 +423,7 @@ class DataLogReader(ResultsReader):
         return url, error_msg
 
 
-class Grinder(RequiredTool):
+class Grinder(RequiredTool):        # todo: take it from maven and convert to JarTool(?)
     def __init__(self, tool_path, parent_logger, version, download_link):
         super(Grinder, self).__init__("Grinder", tool_path, download_link=download_link)
         self.log = parent_logger.getChild(self.__class__.__name__)
